Format deprecation entries in build operation trace
The build operation trace uses the Groovy serializer for formatting entries. Since the deprecation warning operations contain the stacktrace of the callsites in a StackTraceElement list format, the serialization outputs a tremendous amount of unreadable content to the trace file. To solve that, this commit adds custom serialization logic for deprecation progress events.